+++ packages/avatars/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.5](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-avatars@4.0.4...@zendeskgarden/react-avatars@4.0.5) (2018-12-21)
+
+**Note:** Version bump only for package @zendeskgarden/react-avatars
+
+
+
+
+
 ## [4.0.4](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-avatars@4.0.3...@zendeskgarden/react-avatars@4.0.4) (2018-12-10)