+++ packages/menus/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.1](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-menus@4.3.0...@zendeskgarden/react-menus@4.3.1) (2018-12-21)
+
+**Note:** Version bump only for package @zendeskgarden/react-menus
+
+
+
+
+
 # [4.3.0](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-menus@4.2.1...@zendeskgarden/react-menus@4.3.0) (2018-12-19)