+++ packages/tooltips/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.2](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tooltips@5.0.1...@zendeskgarden/react-tooltips@5.0.2) (2018-12-21)
+
+**Note:** Version bump only for package @zendeskgarden/react-tooltips
+
+
+
+
+
 ## [5.0.1](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tooltips@5.0.0...@zendeskgarden/react-tooltips@5.0.1) (2018-12-10)